requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Authlib==1.4.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
deprecation==2.1.0
Flask==3.1.0
Flask-Session==0.8.0
fonttools==4.55.3
frozenlist==1.5.0
gotrue==2.11.2
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
msgspec==0.19.0
multidict==6.1.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
postgrest==0.19.3
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
realtime==2.2.0
redis==5.2.1
requests==2.32.3
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
storage3==0.11.1
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
